updated windows build script, and fixed multiple windows type warnings

+++ HayesDistortion/Source/DSP/WidthProcessor.cpp
@@ -5,15 +5,15 @@ void WidthProcessor::process(float* channeldataL, float* channeldataR, float wid
     // width
     for (int i = 0; i < numSamples; ++i)
     {
-        float mid = (channeldataL[i] + channeldataR[i]) / sqrt(2);   // obtain mid-signal from left and right
-        float sides = (channeldataL[i] - channeldataR[i]) / sqrt(2);   // obtain side-signal from left and right
+        float mid   = (channeldataL[i] + channeldataR[i]) / static_cast<float>(sqrt(2)); // obtain mid-signal from left and right
+        float sides = (channeldataL[i] - channeldataR[i]) / static_cast<float>(sqrt(2)); // obtain side-signal from left and right
 
         // amplify mid and side signal seperately:
         mid *= 2.0f * (1.0f - width);
         sides *= 2.0f * width;
 
-        channeldataL[i] = (mid + sides) / sqrt(2);   // obtain left signal from mid and side
-        channeldataR[i] = (mid - sides) / sqrt(2);   // obtain right signal from mid and side
+        channeldataL[i] = (mid + sides) / static_cast<float>(sqrt(2)); // obtain left signal from mid and side
+        channeldataR[i] = (mid - sides) / static_cast<float>(sqrt(2)); // obtain right signal from mid and side
     }
 }
